Remove some shit from this disgusting heap of shit
