+++ src/main/java/com/pracownia/vanet/model/Crossing.java
@@ -1,39 +1,42 @@
-package com.pracownia.vanet.old;
+package com.pracownia.vanet.model;
 
-import lombok.Data;
+import com.pracownia.vanet.model.point.Point;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import java.util.Date;
 import java.util.Random;
 
-@Data
+@Getter
+@Setter
+@NoArgsConstructor
 public class Crossing {
 
-    public static double DETECTION_RANGE = 1.0;
+    /*------------------------ FIELDS REGION ------------------------*/
+    public static final double DETECTION_RANGE = 1.0;
 
     private Point location;
     private Route routeA;
     private Route routeB;
     private Vehicle lastTransportedVehicle = new Vehicle();
 
-    public Crossing() {
-
-    }
-
+    /*------------------------ METHODS REGION ------------------------*/
     public Crossing(Point location, Route routeA, Route routeB) {
-
         this.location = location;
         this.routeA = routeA;
         this.routeB = routeB;
-
     }
 
     public void transportVehicle(Vehicle vehicle) {
         if (vehicle == lastTransportedVehicle) {
             return;
         }
+
         lastTransportedVehicle = vehicle;
         Random random = new Random();
         int pom = random.nextInt();
+
         if (Math.abs(pom % 3) == 0 || Math.abs(pom % 3) == 1) {
             if (vehicle.getRoute() == routeA) {
                 vehicle.setRoute(routeB);
@@ -68,7 +71,7 @@ public class Crossing {
 
             if (Math.abs(v - vehicle.getSpeed()) > 0.5) {
                 vehicle.setNotSafe("Identified as attacker!");
-            } else if (vehicle.trustLevel < 0.3) {
+            } else if (vehicle.getTrustLevel() < 0.3) {
                 vehicle.setNotSafe("Identified as attacker!");
             } else {
                 //System.out.println("Bezpiecznie.");
@@ -89,3 +92,4 @@ public class Crossing {
         }
     }
 }
+
