+++ src/main/java/com/pracownia/vanet/model/event/EventType.java
@@ -1,4 +1,4 @@
-package com.pracownia.vanet.old;
+package com.pracownia.vanet.model.event;
 
 public enum EventType {
     CAR_ACCIDENT,
