+++ src/main/java/com/pracownia/vanet/model/point/Point.java
@@ -1,0 +1,59 @@
+package com.pracownia.vanet.model.point;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.apache.commons.lang3.builder.EqualsBuilder;
+import org.apache.commons.lang3.builder.HashCodeBuilder;
+import org.apache.commons.lang3.builder.ToStringBuilder;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class Point {
+
+    /*------------------------ FIELDS REGION ------------------------*/
+    private double x = 0.0;
+    private double y = 0.0;
+
+    /*------------------------ METHODS REGION ------------------------*/
+    public Point(double x, double y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+
+        Point point = (Point) o;
+
+        return new EqualsBuilder()
+                .append(x, point.x)
+                .append(y, point.y)
+                .isEquals();
+    }
+
+    @Override
+    public int hashCode() {
+        return new HashCodeBuilder(17, 37)
+                .append(x)
+                .append(y)
+                .toHashCode();
+    }
+
+    @Override
+    public String toString() {
+        return new ToStringBuilder(this)
+                .append("x", x)
+                .append("y", y)
+                .toString();
+    }
+}
+
