--- src/main/java/com/pracownia/vanet/old/App.java
@@ -1,8 +1,0 @@
-package com.pracownia.vanet.old;
-
-public class App {
-    public static void main(String[] args) {
-
-        WindowApp.main(args);
-    }
-}
