--- src/main/java/com/pracownia/vanet/old/Event.java
@@ -1,13 +1,0 @@
-package com.pracownia.vanet.old;
-
-import lombok.Data;
-
-import java.util.Date;
-
-@Data
-public class Event {
-    private int id;
-    private EventType eventType;
-    private Date eventDate;
-    private String message;
-}
