+++ src/main/java/com/pracownia/vanet/util/Logger.java
@@ -1,4 +1,4 @@
-package com.pracownia.vanet.old;
+package com.pracownia.vanet.util;
 
 import java.io.File;
 import java.io.FileWriter;
@@ -8,13 +8,15 @@ import java.text.SimpleDateFormat;
 import java.util.Date;
 
 public class Logger {
+
+    /*------------------------ FIELDS REGION ------------------------*/
     private static Timestamp timeStamp = new Timestamp(System.currentTimeMillis());
-    private static String path =
-            "Log_" + new SimpleDateFormat("HH_mm_ss_dd_MM_yyyy").format(new Date());
+    private static String path = "Log_" + new SimpleDateFormat("HH_mm_ss_dd_MM_yyyy")
+            .format(new Date());
     private static File file = new File(path + ".txt");
 
+    /*------------------------ METHODS REGION ------------------------*/
     public static void log(String msg) {
-
         FileWriter fr = null;
         try {
             fr = new FileWriter(file, true);
@@ -25,3 +27,4 @@ public class Logger {
         }
     }
 }
+
